+++ examples/gplot.py
@@ -17,16 +17,18 @@ dvo=ds(project="EM",simulation="PRE6CPLCr2alb", variable="vo", period="199807", 
 # How to get the required file for rotating vectors from model grid on geographic grid
 fixed_fields('plot', ('angles.nc',cpath+"/../tools/angle_${project}.nc"))
 
-# A Map with one field and vectors, with contours lines like color fill, rotation of vectors on geographic grid
-# and with default projection (a cylindrical equidistant)
+# A Map with one field and vectors, with contours lines like color fill, rotation of vectors on geographic grid,
+# default projection (a cylindrical equidistant), with 'pdf' output format and paper resolution of 17x22 inches (<=> 1224x1584 pixels)
 plot_map1=plot(tos, None, duo, dvo, title='1 field (contours lines follow color filled contours) + vectors',
-              contours=1, rotation=1, vcRefLengthF=0.002, vcRefMagnitudeF=0.02)
-
+              contours=1, rotation=1, vcRefLengthF=0.002, vcRefMagnitudeF=0.02, format="pdf", resolution='17*22') 
+ 
 # Displaying a figure object will compute and cache it if not already done
 cshow(plot_map1)
+# 'cpdfcrop' operator applied on 'plot_map1' object ('cpdfcrop' <=> 'pdfcrop' by preserving metadata)
+cshow(cpdfcrop(plot_map1))
 
-# A Map of one field and vectors, with user-controled contours lines, rotation of vectors on geographic grid
-# and with stereopolar projection
+# A Map of one field and vectors, with user-controled contours lines, rotation as above, stereopolar projection
+# and with 'png' output format (default)
 plot_map2=plot(tos, None, duo, dvo, title='1 field (user-controled contours) + vectors', contours='1 3 5 7 9 11 13',
                proj='NH', rotation=1, vcRefLengthF=0.002, vcRefMagnitudeF=0.02)
 cshow(plot_map2)
@@ -67,7 +69,7 @@ dvo=ds(project="EM",simulation="PRE6CPLCr2alb", variable="vo", period="1998", re
 tos=ds(project="EM",simulation="PRE6CPLCr2alb", variable="tos", period="1998", realm="O") # tos(time_counter, y, x) 
 sub_tos=llbox(tos, latmin=30, latmax=80, lonmin=-60, lonmax=0) # extraction of 'tos' sub box for auxiliary field
 
-map_select2=plot(tos, sub_tos, duo, dvo, title='Selecting level and time closed to 10 and 1400000 respectively', rotation=1, vcRefLengthF=0.002, vcRefMagnitudeF=0.02, level=10., time=1400000.)  # level selection has no impact on main field and auxiliary field because they have not depth dimension, whereas level (and time) selection is done for vectors 
+map_select2=plot(tos, sub_tos, duo, dvo, title='Selecting level and time close to 10 and 1400000 respectively', rotation=1, vcRefLengthF=0.002, vcRefMagnitudeF=0.02, level=10., time=1400000.)  # level selection has no impact on main field and auxiliary field because they have not depth dimension, whereas level (and time) selection is done for vectors 
 cshow(map_select2)